Leonard prime (1 Testcase left)

+++ src/mathprob/leonardPrime.java
@@ -15,7 +15,7 @@ public class leonardPrime {
     static int primeCount(long n) {
 
         int count = 0;
-        long product = 1;
+        double product = 1;
         int[] prime = {2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53};
         for(int i=0; i<prime.length; i++){
             product *= prime[i];
